Update dev-west-2.tfvars
changed vpc_id back as i realized that with each region. whether it's us-east-1. or us-west-2 the default vpc id is gonna be. different, didn't realize thats why when. I tried to. launch it. at first I. was. doing so in. US-east-1 and it had a different vpc_id than the one i had saved it in. originally because I had saved it using us-west-2 vpc_id 😅